requirements.txt
newsapi-python
requests
pandas
ipdb
webhoseio
email
